Find the last line and allow in-range cops.
Many cops, like some of the Metrics cops, analyze a range of lines (an
entire method or class), and if the range of lines isn't paid attention
to, then this gem won't allow that cop to do its thing unless the very
first line in the range (`def method` or `class Foo`) has been touched.

The trick in this code is accurately, 100% identifying the last line of
the range. The implementation I have in here looks like a good attempt,
but time will tell if this works all the time.