Only scan changed files. HT: Dirtycop
This gem grew out of a production setup that was already selecting the
files that had changed on a branch and was feeding that list to
Rubocop, so I'd so far skipped over having Lineup do this itself. But
it should do it itself.

Hat Tip to the dirtycop gem, which unfortunately appears to be
unmaintained, for the tip on where to duck punch Rubocop.

I also had to overhaul this code to retain the full path since Rubocop
works with the full path, and previously I'd simply stripped out the
path from Rubocop when looking up source files in the DiffLiner. But to
achieve this next bit of functionality, I needed to keep the full path,
and so reversed direction there.