0.4.0 - back out "block" support.
Some cops only work when taking into account the outer block of code
involving a changed line, like many of the Metrics cops. Those cops
report their `line_number` as the first line in the block. If anything
inside the block was changed other than the starting line, then the
simple "was this line changed" evaluation keeps it from working.

However, many Style cops work on a line-by-line basis, and really large
blocks (like may exist in a Rails routes.rb file) can have their code
thrashed with auto-corrections that well outnumber the line(s) changed,
which goes against the whole point of this gem.

We'll need to find a better approach to allowing cops to work on blocks
of code where some of it was changed, but that will take some research.